2FOC current limit bugfix, torque control 3 Hz low pass filter.
svn path=/trunk/iCub/firmware/emBODY/any-arch/sys/embobj/; revision=22731

+++ emBODY/eBcode/arch-arm/embobj/mc/src/EOaxisController_hid.h
@@ -66,7 +66,7 @@ struct EOaxisController_hid
     int32_t pos_min;
     int32_t pos_max;
     
-    float   err;
+    int32_t err;
 
     int32_t acc_stop_cmode;
     int32_t acc_stop_alarm;
